Move network journals test (not yet implmented).
